evalops-lab/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "evalops-lab"
version = "0.1.0"
description = "Offline-first EvalOps lab for LLM/RAG evaluation suites, baselines, diffs, and CI gates."
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [{name = "Ethan", email = "ethan@example.com"}]
dependencies = [
  "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.0",
  "ruff>=0.4.0",
]

[project.scripts]
evalops = "evalops_lab.cli:main"

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E","F","I","B","UP"]
ignore = ["E203","E266","E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
